test: reorder test code in a more consistent way
