+++ test/test_api_dev_server.py
@@ -4,7 +4,6 @@ import shutil
 import tempfile
 import unittest
 import uuid
-from unittest import mock
 
 from werkzeug.test import Client
 from werkzeug.wrappers import BaseResponse
@@ -12,7 +11,7 @@ from werkzeug.wrappers import BaseResponse
 from .utils import capture
 
 
-class StandaloneWebApplicationStartupTestCase(unittest.TestCase):
+class DevServerStartupTestCase(unittest.TestCase):
     def setUp(self):
         self.current_path = os.path.dirname(os.path.realpath(__file__))
         self.tempdir = tempfile.mkdtemp()
@@ -53,7 +52,7 @@ class StandaloneWebApplicationStartupTestCase(unittest.TestCase):
         self.assertTrue(os.path.isfile(os.path.join(self.tempdir, "music", "abc.txt")))
 
 
-class StandaloneWebApplicationTestCase(unittest.TestCase):
+class DevServerTestCase(unittest.TestCase):
     def setUp(self):
         from klangbecken.api import development_server
         from klangbecken.cli import init_cmd
@@ -123,17 +122,6 @@ class StandaloneWebApplicationTestCase(unittest.TestCase):
         resp = self.client.get("/data/music/" + fileId + ".flac")
         self.assertEqual(resp.status_code, 200)
         resp.close()
-        #
-        # # Put file in prio list
-        # resp = self.client.post(
-        #     "/api/player/queue/",
-        #     data=json.dumps({"filename": "music/" + fileId + ".flac"}),
-        #     content_type="text/json",
-        #     headers=[("Authorization", f"Bearer {token}")],
-        # )
-        # print(resp.data)
-        # self.assertEqual(resp.status_code, 200)
-        # resp.close()
 
         # Get index.json
         resp = self.client.get("/data/index.json")
@@ -152,83 +140,3 @@ class StandaloneWebApplicationTestCase(unittest.TestCase):
         resp = self.client.post("/api/playlist/music/")
         self.assertEqual(resp.status_code, 401)
         resp.close()
-
-
-class DataDirCreatorTestCase(unittest.TestCase):
-    def setUp(self):
-        self.tempdir = tempfile.mkdtemp()
-
-    def tearDown(self):
-        shutil.rmtree(self.tempdir)
-
-    def testDataDirCheckOnly(self):
-        from klangbecken.settings import PLAYLISTS
-        from klangbecken.utils import _check_data_dir
-
-        for playlist in PLAYLISTS + ("log",):
-            path = os.path.join(self.tempdir, playlist)
-            with self.assertRaises(Exception) as cm:
-                _check_data_dir(self.tempdir, False)
-            self.assertIn("Directory", cm.exception.args[0])
-            self.assertIn("does not exist", cm.exception.args[0])
-            os.mkdir(path)
-
-        for playlist in PLAYLISTS:
-            path = os.path.join(self.tempdir, playlist + ".m3u")
-            with self.assertRaises(Exception) as cm:
-                _check_data_dir(self.tempdir, False)
-            self.assertIn("Playlist", cm.exception.args[0])
-            self.assertIn("does not exist", cm.exception.args[0])
-            with open(path, "a"):
-                pass
-
-        with self.assertRaises(Exception) as cm:
-            _check_data_dir(self.tempdir, False)
-        self.assertIn("File", cm.exception.args[0])
-        self.assertIn("does not exist", cm.exception.args[0])
-
-        with open(os.path.join(self.tempdir, "index.json"), "w"):
-            pass
-
-        _check_data_dir(self.tempdir, False)
-
-    def testDataDirCreation(self):
-        from klangbecken.settings import PLAYLISTS
-        from klangbecken.utils import _check_data_dir
-
-        _check_data_dir(self.tempdir, create=True)
-        for playlist in PLAYLISTS:
-            path = os.path.join(self.tempdir, playlist)
-            self.assertTrue(os.path.isdir(path))
-            path += ".m3u"
-            self.assertTrue(os.path.isfile(path))
-
-        path = os.path.join(self.tempdir, "index.json")
-        self.assertTrue(os.path.isfile(path))
-        with open(path) as f:
-            self.assertEqual(json.load(f), {})
-
-    def testInitCmd(self):
-        from klangbecken.cli import main
-
-        path = os.path.join(self.tempdir, "data")
-        with mock.patch("sys.argv", f"klangbecken init -d {path}".split()):
-            main()
-        self.assertTrue(os.path.exists(path))
-
-        with open(os.path.join(path, "file"), "w"):
-            pass
-
-        with self.assertRaises(SystemExit):
-            with mock.patch("sys.argv", f"klangbecken init -d {path}".split()):
-                with capture(main) as (out, err, ret):
-                    self.assertIn("ERROR", err)
-                    self.assertIn("not empty", err)
-
-        path = os.path.join(path, "file")
-
-        with self.assertRaises(SystemExit):
-            with mock.patch("sys.argv", f"klangbecken init -d {path}".split()):
-                with capture(main) as (out, err, ret):
-                    self.assertIn("ERROR", err)
-                    self.assertIn("not a directory", err)
